move expval test cases into a separate file
